+++ src/dbally/gradio/__init__.py
@@ -1,3 +1,3 @@
-from dbally.gradio.gradio_interface import create_gradio_interface
+from dbally.gradio.interface import create_gradio_interface
 
 __all__ = ["create_gradio_interface"]
